Use join-based spell orphan cleanup

+++ ScriptsDB/20260511-02-remove orphan spell rows.sql
@@ -1,6 +1,1 @@
-DELETE FROM spell
-WHERE NOT EXISTS (
-    SELECT 1
-    FROM "user" u
-    WHERE u.id = spell.user_id
-);
+DELETE FROM spell WHERE user_id IN (SELECT s.user_id FROM spell s LEFT JOIN "user" u ON u.id = s.user_id WHERE u.id IS NULL);
